Enhanced Course Merge
When a user loads a second course into a vehicle this will attempt to detect an intersection between the loaded course and the new course.  If found, it will then create a smooth curve between the two courses.  Crossing points will be ignored.
If not found, it will not append the second course.

The search for an intersection will be performed as follows:
    Start at the "current waypoint" of the already loaded course (course1) until 30 meters after the last waypoint.  The "current waypoint" is the one displayed in the courseplay hud.
        If the courses intersect at multiple locations, the "current waypoint" can be used to designate which intersection to use.  Drive along the loaded course until you get a couple of waypoints before the intersection you want to use.
    Start at a point 30 meters before the first waypoint of the course to load (course2) until the last waypoint.
    The new "Course merge turn radius" will be used at the time of the course merge to create the curve with the designated radius.
        Each loaded course can have a different radius by changing this before the next course is loaded.
    Ctrl-Keypad-5 Increase the Merge Turn Radius
    Ctrl-Keypad-4 Increase the Merge Turn Radius

The usage of this feature is the same as you would use with the crossing points.  The advantage with this method is that the wrong crossing point wont be selected causing strange pathing and a smooth curve is used.

On a map I will create multiple continuous recorded courses for each road/path on the map.  When I need to create a course for a specific task, like a grain sell run, I will load the courses needed and allow the intersections to link the courses.  The biggest issue is with courses that intersect multiple times.  To handle this I will enable course play to drive along the loaded course until I get before the point I want the next intersection to be merged at.  When the crossing course is loaded it will intersect after that point.